Fixed issue with date masking in the frontend. Updated database scripts.
There was an issue with the date masking in the frontend which has beend fixed. Dates were not beign treated as if they were given in UTC, but rather in the system timezone. This fix forces given dates to be interpreted as UTC.
Fixed an issue in a trigger of the database scripts.
The data generation script for the test table now can add null birthdates.